rename kinesis_output to kinesis
